[Changed] from calendar component to calendar hook
